Move register to register folder
